refactor: Refactor deck function
